(PCP-276) Don't use ProcessTerminate in pxp-agent nssm
Using only CTRL-C, not TerminateProcess, to avoid killing the entire
precess tree. Setting the AppKillProcessTree flag to 0.

With this commit, we configure the pxp-agent service for nssm with the
above settings, to prevent killing the entire pxp-agent process tree
once a "service stop" is triggered. Note that the AppKillProcessTree
alone is not sufficient to avoid terminating the entire process tree.